requirements.txt
django==4.1.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
pytest-django==4.5.2
pytest-cov==4.0.0
pytest==7.2.1
requests==2.28.2
gunicorn==20.1.0
pyyaml==6.0
uritemplate==4.1.1
django-rest-swagger==2.2.0
pandas==1.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
boto3==1.26.51
prettyconf==2.2.1
git+https://github.com/olist/olist-loafer@3.0.11
belogging==0.1.3
pydantic==1.10.4
aiohttp==3.8.3
localstack==1.3.1